+++ test/guards/public_files_guard.dart
@@ -1,4 +1,3 @@
-import 'package:dio/dio.dart';
 import 'package:palace/palace.dart';
 import 'package:palace/src/guards/static_file_guard.dart';
 import 'package:test/test.dart';
@@ -6,7 +5,7 @@ import 'package:test/test.dart';
 void main() {
   late Palace router;
 
-  final _dio = Dio(BaseOptions(baseUrl: 'http://localhost:3000', followRedirects: true, validateStatus: (_) => true));
+  // final _dio = Dio(BaseOptions(baseUrl: 'http://localhost:3000', followRedirects: true, validateStatus: (_) => true));
 
   setUp(() async {
     router = Palace();