+++ lib/src/core/palace.dart
@@ -11,12 +11,16 @@ import 'package:palace/utils/logger.dart';
 
 class Palace {
   /// contains the registered end points form any methods like
-  /// `router.get` or `router.post` and so on
-  final List<EndPoint> _endpoints = [];
+  /// `palace.get` or `palace.post` and so on
+  /// will be added here
+  final _endpoints = <EndPoint>[];
 
   /// contains the registered guards [globally] from the `use` method
-  /// `router.use`
+  /// `palace.use`
   final List<GuardFunc> _globalGuards = [
+    /// if logs enabled the logsGuard will be the first one
+    if (PalaceConfig.enaleLogs) LogsGuard(),
+
     /// body parser is on by default
     BodyParser(),
   ];
@@ -27,8 +31,8 @@ class Palace {
   /// assign `Guard` to work globally `on any request with any method`
   void use(GuardFunc guard) => _globalGuards.add(guard);
 
-  EndPoint? match(String method, String path) {
-    // TODO(2) :: find a Better way this will perform badly when many routers exist 'BigO'
+  EndPoint? findMatch(String method, String path) {
+    // TODO :: find a Better way this will perform badly when many routers exist 'BigO'
     try {
       return _endpoints.firstWhere((e) => e.match(method, path));
     } on StateError {
@@ -42,73 +46,94 @@ class Palace {
     required String method,
     required List<GuardFunc> guards,
   }) {
-    _endpoints.add(EndPoint(
-      path: path,
-      method: method,
-      handler: handler,
-      guards: guards,
-    ));
+    /// TODO :: validate if the provided method is supported
+    _endpoints.add(
+      EndPoint(
+        path: path,
+        method: method,
+        handler: handler,
+        guards: guards,
+      ),
+    );
   }
 
+  /// * register the path to work with any type of methods
   void all(
     String path,
     HandlerFunc handler, {
     List<GuardFunc> guards = const [],
   }) {
-    _endpoints.add(EndPoint(
-      path: path,
-      method: '*',
-      handler: handler,
-      guards: guards,
-    ));
+    _endpoints.add(
+      EndPoint(
+        path: path,
+        method: '*',
+        handler: handler,
+        guards: guards,
+      ),
+    );
   }
 
+  /// * register the path to work with `GET` methods
   void get(
     String path,
     HandlerFunc handler, {
     List<GuardFunc> guards = const [],
   }) =>
-      _endpoints.add(EndPoint(
-        path: path,
-        method: 'GET',
-        handler: handler,
-        guards: guards,
-      ));
+      _endpoints.add(
+        EndPoint(
+          path: path,
+          method: 'GET',
+          handler: handler,
+          guards: guards,
+        ),
+      );
 
+  /// * register the path to work with `POST` methods
   void post(
     String path,
     HandlerFunc handler, {
     List<GuardFunc> guards = const [],
   }) =>
-      _endpoints.add(EndPoint(
-        path: path,
-        method: 'POST',
-        handler: handler,
-        guards: guards,
-      ));
+      _endpoints.add(
+        EndPoint(
+          path: path,
+          method: 'POST',
+          handler: handler,
+          guards: guards,
+        ),
+      );
 
+  /// * register the path to work with `GET` methods
   void put(
     String path,
     HandlerFunc handler, {
     List<GuardFunc> guards = const [],
   }) =>
-      _endpoints.add(EndPoint(
-        path: path,
-        method: 'PUT',
-        handler: handler,
-        guards: guards,
-      ));
+      _endpoints.add(
+        EndPoint(
+          path: path,
+          method: 'PUT',
+          handler: handler,
+          guards: guards,
+        ),
+      );
+
+  /// * register the path to work with `PATCH` methods
   void patch(
     String path,
     HandlerFunc handler, {
     List<GuardFunc> guards = const [],
   }) =>
-      _endpoints.add(EndPoint(
-        path: path,
-        method: 'PATCH',
-        handler: handler,
-        guards: guards,
-      ));
+      _endpoints.add(
+        EndPoint(
+          path: path,
+          method: 'PATCH',
+          handler: handler,
+          guards: guards,
+        ),
+      );
+
+  /// * register the path to work with `DELETE` methods
   void delete(
     String path,
     HandlerFunc handler, {
@@ -121,6 +146,7 @@ class Palace {
         guards: guards,
       ));
 
+  /// * register the path to work with `GET` methods
   void _bootstrap() {
     for (final e in _endpoints) {
       if (_endpoints.where((element) {
@@ -131,28 +157,33 @@ class Palace {
             }
           }).length >
           1) {
+        // TODO :: better error message
         throw 'your endpoints have a duplicate try to fix it => $e';
       }
     }
   }
 
+  /// * open the server and start listening for the incoming requests
   Future<void> openGates({
     int port = 3000,
     String? ip,
   }) async {
+    /// checks for endpoints dublicate
     _bootstrap();
-    // await PalaceDB.init();
+
+    /// await PalaceDB.init();
     ip ??= InternetAddress.anyIPv4.address;
-    if (PalaceConfig.enaleLogs) {
-      _globalGuards.insert(0, LogsGuard());
-    }
+
+    /// incase of old server is on it will forcly close it
     await _server?.close(force: true);
 
-    /// open the server
+    /// create the new server
     _server = await HttpServer.bind(ip, port);
 
-    /// print the server url
-    print('Listening on http://localhost:$port');
+    if (PalaceConfig.enaleLogs) {
+      /// print the server url
+      print('Listening on http://localhost:$port');
+    }
 
     /// add listener to incoming requests stream
     _server!.listen(_mainPipe);
@@ -162,7 +193,8 @@ class Palace {
     /// * wait for incoming requests
     try {
       /// * look for desired endpoint
-      final endpoint = match(ioReq.method, ioReq.uri.path) ??
+      /// if none it will use `notFoundHandler` from `PalaceConfig`
+      final endpoint = findMatch(ioReq.method, ioReq.uri.path) ??
           EndPoint(
             path: ioReq.uri.path,
             method: ioReq.method,
@@ -171,16 +203,19 @@ class Palace {
 
       /// * create Place req form dart io req and the desired endpoint;
       final req = await Request.init(ioReq, endpoint);
+
+      /// * create palace res from the incoming request
       final res = Response(ioReq);
 
-      /// build list of guards
+      /// * build list of guards from the public guards then the request gurads
+      final _reqGuards = <GuardFunc>[
+        ..._globalGuards,
+        ...endpoint.guards,
+      ];
 
-      final _reqGuards = <GuardFunc>[..._globalGuards, ...endpoint.guards];
+      /// idk what i have done here its was 3 month ago so i will leave it like this for now
       final queue = [];
-      // Future<void> handelCall(Function func) async{
-      //   await
 
-      // }
       for (var i = 0; i <= _reqGuards.length; i++) {
         if (i == _reqGuards.length) {
           queue.add(() => endpoint.handler(req, res));
@@ -193,15 +228,17 @@ class Palace {
     } on PalaceResponse catch (e) {
       Response(ioReq).json(e.toMap(), statusCode: e.statusCode);
     } catch (e, st) {
-      await Logger.e(e, st: st);
+      /// if logs enabled log the err
+      if (PalaceConfig.enaleLogs) await Logger.e(e, st: st);
       Response(ioReq).json(InternalServerError(e).toMap());
     } finally {
-      ///  Close the req
+      ///* Close the req
       await ioReq.response.close();
     }
   }
 
   Future<void> _handleResponse(Request req, Response res, dynamic data) async {
+    // TODO :: sever Views in future
     if (data is String) {
       res.send(data);
     } else if (data is Map) {
@@ -210,8 +247,10 @@ class Palace {
       res.json(data.toMap(), statusCode: data.statusCode);
     } else if (data is File) {
       await file(req.request, res.response, data);
+    } else {
+      res.send(data);
     }
   }
 
-  Future<void> closeGates() async => await _server?.close(force: true);
+  Future<dynamic>? closeGates() => _server?.close(force: true);
 }