+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.1+3
 
 - no more `palace.use(BodyParser());` its its powered by `palace_body_parser` out of the box 🎁
+- fix ip bug
 
 ## 0.0.1+2